+++ Demo/iOS/MainMenu.swift
@@ -18,11 +18,11 @@ struct MainMenu: View {
         Menu {
             RichTextShareMenu(
                 formats: .libraryFormats,
-                formatAction:  { print("TODO: Share file for \($0.id)") },
+                formatAction: { print("TODO: Share file for \($0.id)") },
                 pdfAction: { print("TODO: Share PDF file") })
             RichTextExportMenu(
                 formats: .libraryFormats,
-                formatAction:  { print("TODO: Export file for \($0.id)") },
+                formatAction: { print("TODO: Export file for \($0.id)") },
                 pdfAction: { print("TODO: Export PDF file") })
             Divider()
             link(to: .about)