+++ Sources/RichTextKit/Context/RichTextContext.swift
@@ -19,201 +19,164 @@ import SwiftUI
  */
 public class RichTextContext: ObservableObject {
 
-    /**
-     Create a new rich text context.
-     */
+    /// Create a new rich text context instance.
     public init() {}
+    
+    
+    // MARK: - Not yet observable properties
+    
+    /**
+     The currently attributed string, if any.
+     
+     Note that the property is read-only and not `@Published`
+     to avoid redrawing the editor when it changes, which is
+     done constantly as the user types. We should find a way
+     to observe this property without this happening.
+     
+     The best way to observe this property is to use the raw
+     `text` binding that you pass into your text editor. The
+     editor will however not redraw if you change this value
+     from the outside, since it aims to avoid costly redraws.
 
-
+     Until then, use `setAttributedString(to:)` to change it.
+     */
+    public internal(set) var attributedString = NSAttributedString()
+    
     /**
      The currently selected range, if any.
+     
+     Note that the property is read-only and not `@Published`
+     to avoid redrawing the editor when it changes, which is
+     done constantly as the user types. We should find a way
+     to observe this property without this happening.
 
-     Note that this property is currently not @Published, in
-     order to avoid redrawing the entire app. We should find
-     a way to make it published once more without it causing
-     the entire app to redraw every time the cursor is moved.
-     Until then, the property has been made read-only, while
-     a hopefully temporary `selectedRangeChange` property is
-     now used by ``selectRange(_:)``.
+     Until then, use ``selectRange(_:)`` to change the value.
      */
     public internal(set) var selectedRange = NSRange()
 
-    /**
-     The selected range to change to.
-     */
-    @Published
-    var selectedRangeChange = NSRange()
 
-
-    /**
-     The current background color, if any.
-     */
+    // MARK: - Observable properies
+    
+    /// The current background color, if any.
     @Published
     public var backgroundColor: ColorRepresentable?
 
-    /**
-     Whether or not the current rich text can be copied.
-     */
+    /// Whether or not the current rich text can be copied.
     @Published
     public var canCopy = false
 
-    /**
-     Whether or not the latest undone change can be redone.
-     */
+    /// Whether or not the latest undone change can be redone.
     @Published
     public var canRedoLatestChange = false
 
-    /**
-     Whether or not the latest change can be undone.
-     */
+    /// Whether or not the latest change can be undone.
     @Published
     public var canUndoLatestChange = false
     
-    /**
-     Whether or not the indent level can be decreased.
-     */
+    /// Whether or not the indent level can be decreased.
     @Published
     public var canDecreaseIndent = true
     
-    /**
-     Whether or not the indent level can be increased.
-     */
+    /// Whether or not the indent level can be increased.
     @Published
     public var canIncreaseIndent = true
 
-    /**
-     The current font name.
-     */
+    /// The current font name.
     @Published
     public var fontName = ""
 
-    /**
-     The current font size.
-     */
+    /// The current font size.
     @Published
     public var fontSize = CGFloat.standardRichTextFontSize
 
-    /**
-     The current foreground color, if any.
-     */
+    /// The current foreground color, if any.
     @Published
     public var foregroundColor: ColorRepresentable?
 
-    /**
-     The currently highlighted range, if any.
-     */
+    /// The currently highlighted range, if any.
     @Published
     public var highlightedRange: NSRange?
 
-    /**
-     The style to apply when highlighting a range.
-     */
+    /// The style to apply when highlighting a range.
     @Published
     public var highlightingStyle = RichTextHighlightingStyle.standard
 
-    /**
-     Whether or not the current text is bold.
-     */
+    /// Whether or not the current text is bold.
     @Published
     public var isBold = false
 
-    /**
-     Whether or not the rich text is currently being edited.
-     */
+    /// Whether or not the text is currently being edited.
     @Published
     public var isEditingText = false
 
-    /**
-     Whether or not the current text is italic.
-     */
+    /// Whether or not the current text is italic.
     @Published
     public var isItalic = false
 
-    /**
-     Whether or not the current text is striked through.
-     */
+    /// Whether or not the current text is striked through.
     @Published
     public var isStrikethrough = false
 
-    /**
-     Whether or not the current text is underlined.
-     */
+    /// Whether or not the current text is underlined.
     @Published
     public var isUnderlined = false
-
-    /**
-     Whether or not to copy the current text selection.
-     */
+    
+    /// The current text alignment, if any.
+    @Published
+    public var textAlignment: RichTextAlignment = .left
+    
+    
+    // MARK: - Internal trigger properties
+    
+    /// Whether or not to copy the current text selection.
     @Published
     public var shouldCopySelection = false
 
-    /**
-     Whether or not to redo the latest undone change.
-     */
+    /// Whether or not to redo the latest undone change.
     @Published
     public var shouldRedoLatestChange = false
 
-    /**
-     Set this property to trigger an image paste operation.
-     */
+    /// Set this property to trigger a paste operation.
     @Published
     public var shouldPasteImage: (image: ImageRepresentable, atIndex: Int, moveCursor: Bool)?
 
-    /**
-     Set this property to trigger an image paste operation.
-     */
+    /// Set this property to trigger an image paste operation.
     @Published
     public var shouldPasteImages: (images: [ImageRepresentable], atIndex: Int, moveCursor: Bool)?
 
-    /**
-     Set this property to trigger a text paste operation.
-     */
+    /// Set this property to trigger a text paste operation.
     @Published
     public var shouldPasteText: (text: String, atIndex: Int, moveCursor: Bool)?
 
-    /**
-     Set this property to trigger an attributed string update.
-     */
+    /// Set this property to trigger a string update.
     @Published
     public var shouldSetAttributedString: NSAttributedString?
+    
+    /// Set this property to trigger a range change.
+    @Published
+    var shouldSelectRange = NSRange()
 
-    /**
-     Whether or not to undo the latest change.
-     */
+    /// Whether or not to undo the latest change.
     @Published
     public var shouldUndoLatestChange = false
 
-    /**
-     Whether or not to decrease the current indent.
-     */
+    /// Whether or not to decrease the current indent.
     @Published
     public var shouldDecreaseIndent = false
     
-    /**
-     Whether or not to increase the current indent.
-     */
+    /// Whether or not to increase the current indent.
     @Published
     public var shouldIncreaseIndent = false
-    
-    /**
-     The current text alignment, if any.
-     */
-    @Published
-    public var textAlignment: RichTextAlignment = .left
 }
 
 public extension RichTextContext {
 
-    /**
-     Whether or not the context has a selected range.
-     */
+    /// Whether or not the context has a selected range.
     var hasHighlightedRange: Bool {
         highlightedRange != nil
     }
 
-    /**
-     Whether or not the context has a selected range.
-     */
+    /// Whether or not the context has a selected range.
     var hasSelectedRange: Bool {
         selectedRange.length > 0
     }
@@ -221,35 +184,22 @@ public extension RichTextContext {
 
 public extension RichTextContext {
 
-    /**
-     Copy the current selection to the pasteboard.
-
-     This is only usable on platforms that have a pasteboard.
-     */
+    /// Copy the current selection to the pasteboard.
     func copyCurrentSelection() {
         shouldCopySelection = true
     }
 
-    /**
-     Decrement the current font size 1 point.
-     */
+    /// Decrement the current font size 1 point.
     func decrementFontSize() {
         decrementFontSize(points: 1)
     }
 
-    /**
-     Decrement the current font size.
-
-     - Parameters:
-       - points: The number of points to decrement the font size.
-     */
+    /// Decrement the current font size a number of points.
     func decrementFontSize(points: UInt) {
         stepFontSize(points: -Int(points))
     }
 
-    /**
-     Whether or not a certain style is enabled.
-     */
+    /// Whether or not a certain style is enabled.
     func hasStyle(_ style: RichTextStyle) -> Bool {
         switch style {
         case .bold: return isBold
@@ -259,26 +209,17 @@ public extension RichTextContext {
         }
     }
 
-    /**
-     Set ``highlightedRange`` to a new, optional range.
-     */
+    /// Set ``highlightedRange`` to a new, optional range.
     func highlightRange(_ range: NSRange?) {
         highlightedRange = range
     }
 
-    /**
-     Increment the current font size 1 point.
-     */
+    /// Increment the current font size 1 point.
     func incrementFontSize() {
         incrementFontSize(points: 1)
     }
 
-    /**
-     Increment the current font size.
-
-     - Parameters:
-       - points: The number of points to increment the font size.
-     */
+    /// Increment  the current font size a number of points.
     func incrementFontSize(points: UInt) {
         stepFontSize(points: Int(points))
     }
@@ -331,88 +272,56 @@ public extension RichTextContext {
         shouldPasteText = (text, index, moveCursorToPastedContent)
     }
 
-    /**
-     Redo the latest undone change.
-     */
+    /// Redo the latest undone change.
     func redoLatestChange() {
         shouldRedoLatestChange = true
     }
 
-    /**
-     Reset the attributed string.
-     */
+    /// Reset the attributed string.
     func resetAttributedString() {
         setAttributedString(to: "")
     }
 
-    /**
-     Reset the ``highlightedRange``.
-     */
+    /// Reset the ``highlightedRange``.
     func resetHighlightedRange() {
         guard hasHighlightedRange else { return }
         highlightedRange = nil
     }
 
-    /**
-     Reset the ``selectedRange``.
-     */
+    /// Reset the ``selectedRange``.
     func resetSelectedRange() {
         selectedRange = NSRange(location: 0, length: 0)
     }
 
-    /**
-     Set ``selectedRange`` to a new range and start editing.
-
-     - Parameters:
-       - range: The range to select.
-     */
+    /// Set a new range and start editing.
     func selectRange(_ range: NSRange) {
         isEditingText = true
-        selectedRangeChange = range
+        shouldSelectRange = range
     }
 
-    /**
-     Set the attributed string to a new plain text.
-
-     - Parameters:
-       - text: The plain text string to set.
-     */
+    /// Set the attributed string to a new plain text.
     func setAttributedString(to text: String) {
         setAttributedString(to: NSAttributedString(string: text))
     }
 
-    /**
-     Set the attributed string to a new rich text.
-
-     - Parameters:
-       - string: The rich text string to set.
-     */
+    /// Set the attributed string to a new rich text.
     func setAttributedString(to string: NSAttributedString) {
         let mutable = NSMutableAttributedString(attributedString: string)
         mutable.setFontSize(to: fontSize)
         shouldSetAttributedString = mutable
     }
 
-    /**
-     Step the current font size a certain number of points.
-
-     - Parameters:
-       - points: The number of points to increase or decrease the font size.
-     */
+    /// Step the current font size a number of points.
     func stepFontSize(points: Int) {
         fontSize += CGFloat(points)
     }
 
-    /**
-     Set ``isEditingText`` to `false`.
-     */
+    /// Set ``isEditingText`` to `false`.
     func stopEditingText() {
         isEditingText = false
     }
 
-    /**
-     Toggle a certain rich text style.
-     */
+    /// Toggle a certain rich text style.
     func toggle(_ style: RichTextStyle) {
         switch style {
         case .bold: isBold.toggle()
@@ -422,30 +331,22 @@ public extension RichTextContext {
         }
     }
 
-    /**
-     Toggle whether or not the text is being edited.
-     */
+    /// Toggle whether or not the text is being edited.
     func toggleIsEditing() {
         isEditingText.toggle()
     }
 
-    /**
-     Undo the latest change.
-     */
+    /// Undo the latest change.
     func undoLatestChange() {
         shouldUndoLatestChange = true
     }
     
-    /**
-     Decrease current indent.
-     */
+    /// Decrease current indent.
     func decreaseIndent() {
         shouldDecreaseIndent = true
     }
     
-    /**
-     Increase current indent.
-     */
+    /// Increase current indent.
     func increaseIndent() {
         shouldIncreaseIndent = true
     }